varlink_derive: use path relative to manifest dir
Currently, varlink_file! macro takes as argument either an absolute path
or a relative path to the current working directory. This can be
different depending on the Rust/Cargo version.

The documentation says it is relative to the workspace directory, but
that isn't true for the MSRV 1.70.0.

This commit inspects the env var CARGO_MANIFEST_DIR, if it is defined,
and prepends it to the given filename to create an absolute path. This
way, the behavior of the macro is consistent.

Signed-off-by: Manos Pitsidianakis <manos@pitsidianak.is>